fix(web): move notifications bell into the home domain; amber top-left dot
Codex review: shared components/ must not import domain state or feature
hooks (STYLE_GUIDE — Shared UI components), and the bell pulls the home
feed. It now lives in domains/home/components/, self-contained via the
resolved-assistants store, and routes.tsx injects it into ChatLayout as a
topBarAccessory prop — neither the chat nor home domain imports the other.

Also restyles the unread dot to match the amber top-left dots on unread
rows inside the popover (was red top-right).
